+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 3.15.23 (February 1, 2023)
+
+### Added
+
+- Add subnav parent icon to Subnav component
+
+### Fixed
+
+- Fix prevent background scrolling in Safari
+
 ## 3.15.22 (January 19, 2023)
 
 ### Added